refactor: unify mock_pma_entry implementations
